requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-hashes --no-dev
-e .
certifi==2025.1.31
cffi==1.17.1 ; python_full_version >= '3.13'
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
dlt==1.9.0
dlt-pendulum==3.0.2 ; python_full_version >= '3.13'
duckdb==1.2.1
fsspec==2025.3.1
gitdb==4.0.12
gitpython==3.1.44
giturlparse==0.12.0
hexbytes==1.3.0
humanize==4.12.2
idna==3.10
jsonpath-ng==1.7.0
makefun==1.15.6
markdown-it-py==3.0.0
mdurl==0.1.2
orjson==3.10.16 ; platform_python_implementation != 'PyPy'
packaging==24.2
pathvalidate==3.2.3
pendulum==3.0.0 ; python_full_version < '3.13'
pluggy==1.5.0
ply==3.11
polars==1.26.0
pycparser==2.22 ; python_full_version >= '3.13'
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310 ; sys_platform == 'win32'
pyyaml==6.0.2
requests==2.32.3
requirements-parser==0.11.0
rich==14.0.0
rich-argparse==1.7.0
semver==3.0.4
setuptools==78.1.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sqlglot==26.12.0
tenacity==9.0.0
time-machine==2.16.0 ; python_full_version < '3.13' and implementation_name != 'pypy'
tomlkit==0.13.2
types-setuptools==78.1.0.20250329
typing-extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
win-precise-time==1.4.2 ; os_name == 'nt'
